Set catch-all server name
The goal of this change is to mask the location of the ELB that sets
its IP as the Host when forwarding requests. This results in nginx
redirects containing the IP address of the ELB. The intention is to
set the server_name so it can get set in the Host header of the
redirect response.

Used documentation here:
http://nginx.org/en/docs/http/server_names.html
http://nginx.org/en/docs/http/request_processing.html